% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clio_get.R
\name{clio_get}
\alias{clio_get}
\title{Execute a data query to clio-infra}
\usage{
clio_get(variables, countries, from, to, list = FALSE, mergetype = full_join)
}
\arguments{
\item{variables}{The variables you want to obtain (provided to you by \code{\link[Clio:clio_overview]{Clio::clio_overview()}})}

\item{countries}{If left empty, all countries.}

\item{from}{Start year}

\item{to}{End year}

\item{list}{Defaults to FALSE. If TRUE, returns a list of each variable}

\item{mergetype}{Defaults to full_join. Can be set to inner_join, outer_join, left_join, etc.}
}
\value{
a data.frame containing the required variables,
countries and years specified by the user.
}
\description{
This function is used to execute a query to clio-infra,
and returns a data.frame containing the required variables,
countries and years specified by the user.
}
\examples{
clio_get(c("infant mortality", "zinc production"))

clio_get(c("biodiversity - naturalness", "xecutive Constraints  (XCONST)"),
from = 1850, to = 1900,
countries = c("Armenia", "Azerbaijan"))

clio_get(c("Zinc production", "Gold production"),
from = 1800, to = 1920,
countries = c("Botswana", "Zimbabwe",
             mergetype = inner_join))
}
\seealso{
\code{\link[Clio:clio_get_cat]{Clio::clio_get_cat()}} if you want to extract data by categories instead of by variables.
}
