% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clio_get_cat.R
\name{clio_get_cat}
\alias{clio_get_cat}
\title{Executive a query of categories to clio-infra}
\usage{
clio_get_cat(category, ...)
}
\arguments{
\item{category}{A vector of categories.}

\item{...}{All other arguments are passed on to \code{\link[Clio:clio_get]{Clio::clio_get()}}}
}
\value{
A data.frame (or list) consisting of all the variables within a certain category
}
\description{
This function allows you to enter (one or more of the available) category
names as an arguments.
}
\examples{

clio_get_cat("finanz", list = F, from = 1800, to = 1900)

clio_get_cat(c("agriculture", "environment"),
countries = "Netherlands",
mergetype = inner_join, from = 1850, to = 1900)
clio_get_cat("Produzioni", from = 1700, list = F, mergetype = inner_join)

clio_get(c("Tin Production", "income inequality"), from = 1800, countries = c("Netherlands", "Russia"))

clio_get_cat("labor relation")
}
\seealso{
\code{\link[Clio:clio_get]{Clio::clio_get()}} Arguments other than category are passed to this function.
}
