% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clio_overview.R
\name{clio_overview}
\alias{clio_overview}
\title{Provide an overview of all available variables on Clio-infra}
\usage{
clio_overview()
}
\value{
A data.frame with the variable names, from, to and number of observations (obs)
}
\description{
The function returns a data.frame containing all variables, and indicates availability
in terms of years and observations.
}
\examples{
clio_overview()
}
\seealso{
\code{\link[Clio:clio_overview_cat]{Clio::clio_overview_cat()}}
}
