% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clio_overview_cat.R
\name{clio_overview_cat}
\alias{clio_overview_cat}
\title{Overview of categories or variables within a category}
\usage{
clio_overview_cat(subargument = "all")
}
\arguments{
\item{subargument}{One of the categories present on Clio-infra}
}
\value{
With no argument, it returns a vector containing the names of all available categories
on clio-infra. With a category as one of its arguments, it returns the names, from, to and obs of
all variables within that category.
}
\description{
This function returns a vector of categories or all variables within a certain category
(indicated by an argument)
}
\examples{
clio_overview_cat("finanz")
clio_overview_cat()
clio_overview_cat("institutions")
}
