% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{data_clean}
\alias{data_clean}
\title{Auxilary function to help clean the raw data}
\usage{
data_clean(x)
}
\value{
A data.frame, but for internal use only.
}
\description{
Not to be used seperately. Executes a few tidyverse commands
to quickly and efficiently clean the data for user convenience.
}
